+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.0.x / 2013-03-25
+------------------
+* renamed `Value` to `StringValue`
+* changed from vals to pointers for `Options`
+
 0.0.1 / 2013-03-22 
 ------------------
 * forked from https://github.com/vanetix/commander.go